started from exc need to fix bugs

+++ ft_atoi.c
@@ -6,21 +6,17 @@
 /*   By: keblazer <keblazer@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/29 11:22:33 by keblazer       #+#    #+#                */
-/*   Updated: 2019/11/21 02:18:51 by keblazer      ########   odam.nl         */
+/*   Updated: 2019/11/21 05:32:22 by keblazer      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-//	TODO:
-//	[KO]: your atoi does not work with over long max value
-//	[KO]: your atoi does not work with over long min value
-
 int	ft_atoi(const char *str)
 {
-	int				i;
-	long long int	result;
-	long long int	negative;
+	int	i;
+	int	result;
+	int	negative;
 
 	result = 0;
 	negative = 1;
@@ -38,7 +34,7 @@ int	ft_atoi(const char *str)
 		if (str[i] < 48 || 57 < str[i])
 			return (result * negative);
 		else
-			result = (result * 10) + (long long int)(str[i] - '0');
+			result = (result * 10) + (int)(str[i] - '0');
 		i++;
 	}
 	return (result * negative);
